changing data type
as if you will put number of 6 digits then int  datatype is not ok.

+++ practical-22dec/question1.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main(){
-      int number , lastdigit ;
+      long int number , lastdigit ;
       cout<<"WRITE ANY NUMBER :   \n";
       cin>>number ;
       lastdigit = number%10 ;
